simplify error overlay
In the previous version, we:
1. wrapped the output component with an overlay. This could cause flow
issues by introducing new markup on the page (e.g. wrapping a div with
a button).
2. Disabled the input components

This commit simplifies the logic and just displays the previous state
of the component, removes the overlay, and removes the disabled
elements.

Now, the markup on the page won’t change when a front-end error happens
which makes it more similar to how the app changes when backend errors
occur.